+++ tests/testthat/test-prev-wasting-mfaz.R
@@ -1,5 +1,12 @@
-# Test check: mw_estimate_prevalence_mfaz ----
-## When std != problematic & is.null(.wt) & !is.null(oedema) ----
+# ==============================================================================
+# 📦 Functions: mw_estimate_prevalence_mfaz()
+# ==============================================================================
+
+
+## ---- Test check:  mw_estimate_prevalence_mfaz() -------------------------
+
+
+### When std != problematic & is.null(.wt) & !is.null(oedema) ----
 testthat::test_that(
   "mw_estimate_prevalence_mfaz() yields correct estimates",
   {
@@ -42,7 +49,8 @@ testthat::test_that(
   }
 )
 
-## When std != problematic & is.null(.wt) & is.null(oedema) ----
+
+### When std != problematic & is.null(.wt) & is.null(oedema) ----
 testthat::test_that(
   "mw_estimate_prevalence_mfaz() yields correct estimates",
   {
@@ -86,21 +94,24 @@ testthat::test_that(
 )
 
 
-## When standard deviation == problematic ----
+### When standard deviation == problematic ----
 testthat::test_that(
   "mw_estimate_prevalence_mfaz() works well on a dataframe with multiple survey areas with
     different categories on analysis_approach",
   {
     ### Get the prevalence estimates ----
     p <- anthro.04 |>
-      mw_estimate_prevalence_mfaz(oedema = oedema, wt = NULL, province)
-
-    ### Subset a province whose analysis approach is unweighted ---
-    province_1 <- subset(p, province == "Province 1")
-
-    ### Subset a province whose analysis approach is weighted ---
-    province_3 <- subset(p, province == "Province 3")
-
+      mw_wrangle_age(age = age) |>
+      mw_wrangle_muac(
+        muac = muac,
+        .recode_muac = TRUE,
+        .to = "cm",
+        age = age,
+        sex = sex,
+        .recode_sex = FALSE
+      ) |>
+      transform(muac = recode_muac(muac, "mm")) |>
+      mw_estimate_prevalence_mfaz(oedema = oedema, wt = NULL, analysis_unit)
 
     columns_to_check <- c(
       "gam_n", "gam_p_low", "gam_p_upp", "sam_n",
@@ -109,9 +120,9 @@ testthat::test_that(
     )
 
     ### Tests ----
-    testthat::expect_vector(dplyr::select(p, !province), size = 3, ncol(17))
+    testthat::expect_vector(dplyr::select(p, !analysis_unit), size = 3, ncol(17))
     testthat::expect_s3_class(p, "tbl")
-    testthat::expect_false(all(sapply(province_1[columns_to_check], \(.) all(is.na(.)))))
-    testthat::expect_true(all(sapply(province_3[columns_to_check], \(.) all(is.na(.)))))
+    testthat::expect_false(all(sapply(p[2, ][columns_to_check], \(.) all(is.na(.)))))
+    testthat::expect_true(all(sapply(p[3, ][columns_to_check], \(.) all(is.na(.)))))
   }
 )