+++ tests/testthat/test-prev-wasting-screening.R
@@ -1,9 +1,17 @@
-# Test check: get_estimates() ----
-## When grouping variables are not supplied ----
+# ==============================================================================
+# 📦 Functions: get_estimates(), mw_estimate_prevalence_screening(), and 
+# mw_estimate_prevalence_screening2()
+# ==============================================================================
+
+
+## ---- Test check: get_estimates() --------------------------------------------
+
+
+### When grouping variables are not supplied ----
 testthat::test_that(
   "get_estimates() works OK",
   {
-    ## Wrangle data ----
+    #### Wrangle data ----
     df <- anthro.02 |>
       mutate(
         muacx = as.character(muac),
@@ -11,13 +19,8 @@ testthat::test_that(
         ede = ifelse(oedema == "y", "yes", 0)
       )
 
-    ### Get estimates ----
-    p <- df |>
-      get_estimates(
-        muac = muac,
-        oedema = oedema,
-        raw_muac = FALSE
-      )
+    #### Get estimates ----
+    p <- get_estimates(df, muac, oedema, FALSE)
 
     ### Observed estimates ----
     gam_n <- 118
@@ -27,7 +30,7 @@ testthat::test_that(
     mam_n <- 89
     mam_p <- 4.1
 
-    ### Tests ----
+    #### Tests ----
     testthat::expect_s3_class(p, "tbl_df")
     testthat::expect_equal(ncol(p), 7)
     testthat::expect_equal(nrow(p), 1)
@@ -43,17 +46,12 @@ testthat::test_that(
     testthat::expect_error(
       anthro.02 |>
         mutate(muac = recode_muac(muac, .to = "cm")) |>
-        get_estimates(
-          muac = muac,
-          oedema = oedema
-        ),
+        get_estimates(muac = muac, oedema = oedema),
       regexp = "MUAC values must be in millimetres. Try again!"
     )
     testthat::expect_error(
       df |>
-        get_estimates(
-          muac = muacx,
-          oedema = oedema
+        get_estimates(muac = muacx, oedema = oedema
         ),
       regexp = paste0(
         "`muac` should be of class numeric not ",
@@ -62,9 +60,7 @@ testthat::test_that(
     )
     testthat::expect_error(
       df |>
-        get_estimates(
-          muac = muac,
-          oedema = oedemax
+        get_estimates(muac = muac, oedema = oedemax
         ),
       regexp = paste0(
         "`oedema` should be of class character not ",
@@ -73,27 +69,21 @@ testthat::test_that(
     )
     testthat::expect_error(
       df |>
-        get_estimates(
-          muac = muac,
-          oedema = ede
+        get_estimates(muac = muac, oedema = ede
         ),
       regexp = 'Code values in `oedema` must only be "y" and "n". Try again!'
     )
   }
 )
 
-## When is.null(oedema) & grouping variables are not supplied ----
+### When is.null(oedema) & grouping variables are not supplied ----
 testthat::test_that(
   "get_estimates() works OK when oedema and grouping variables null",
   {
-    ### Get estimates ----
-    p <- anthro.02 |>
-      get_estimates(
-        muac = muac,
-        oedema = NULL
-      )
+    #### Get estimates ----
+    p <- get_estimates(anthro.02, muac, oedema = NULL)
 
-    ### Observed estimates ----
+    #### Observed estimates ----
     gam_n <- 106
     gam_p <- 4.9
     sam_n <- 16
@@ -101,7 +91,7 @@ testthat::test_that(
     mam_n <- 90
     mam_p <- 4.1
 
-    ### Tests ----
+    #### Tests ----
     testthat::expect_s3_class(p, "tbl_df")
     testthat::expect_equal(ncol(p), 7)
     testthat::expect_equal(nrow(p), 1)
@@ -117,20 +107,14 @@ testthat::test_that(
   }
 )
 
-## When grouping variables are supplied ----
+### When grouping variables are supplied ----
 testthat::test_that(
   "get_estimates() works OK when grouping variables are supplied",
   {
-    ### Get estimates ----
-    p <- anthro.02 |>
-      get_estimates(
-        muac = muac,
-        oedema = oedema,
-        raw_muac = FALSE,
-        province
-      )
+    #### Get estimates ----
+    p <- get_estimates(anthro.02, muac, oedema, raw_muac = FALSE, province)
 
-    ### Observed estimates ----
+    #### Observed estimates ----
     gam_n <- 61
     gam_p <- 5.9
     sam_n <- 19
@@ -138,7 +122,7 @@ testthat::test_that(
     mam_n <- 42
     mam_p <- 4.1
 
-    ### Tests ----
+    #### Tests ----
     testthat::expect_s3_class(p, "tbl_df")
     testthat::expect_equal(ncol(p), 8)
     testthat::expect_equal(nrow(p), 2)
@@ -155,37 +139,30 @@ testthat::test_that(
   }
 )
 
-## When `raw_muac` is either `TRUE` or `FALSE` ----
+### When `raw_muac` is either `TRUE` or `FALSE` ----
 testthat::test_that(
   "When get_estimates() is set to `raw_muac = TRUE`, it filters outliers
   based on `flag_muac`",
   {
-    ### Observed results ----
+    #### Observed results ----
     r <- anthro.01 |>
       mw_wrangle_age(age = age) |>
-      mw_wrangle_muac(
-        sex = sex,
-        .recode_sex = TRUE,
-        muac = muac
-      ) |>
-      get_estimates(
-        muac = muac,
-        raw_muac = TRUE
-      )
+      mw_wrangle_muac(sex = sex, .recode_sex = TRUE, muac = muac) |>
+      get_estimates(muac = muac, raw_muac = TRUE)
 
-    ### Tests ----
+    #### Tests ----
     testthat::expect_s3_class(object = r, class = "tbl_df")
     testthat::expect_no_error(object = r)
   }
 )
 
 
-## When `raw_muac` is either `TRUE` or `FALSE` ----
+### When `raw_muac` is either `TRUE` or `FALSE` ----
 testthat::test_that(
   "When get_estimates() is set to `raw_muac = FALSE`, it filters outliers
   based on `flag_mfaz`",
   {
-    ### Observed results ----
+    #### Observed results ----
     r <- anthro.01 |>
       mw_wrangle_age(age = age) |>
       mw_wrangle_muac(
@@ -197,10 +174,7 @@ testthat::test_that(
         .to = "cm"
       ) |>
       mutate(muac = recode_muac(muac, .to = "mm")) |>
-      get_estimates(
-        muac = muac,
-        raw_muac = FALSE
-      )
+      get_estimates(muac = muac, raw_muac = FALSE)
 
     ### Tests ----
     testthat::expect_s3_class(object = r, class = "tbl_df")
@@ -208,56 +182,18 @@ testthat::test_that(
   }
 )
 
-# Test check: mw_estimate_prevalence_screening() ----
-testthat::test_that(
-  "mw_estimate_prevalence_screening() works OK",
-  {
-    ### Get estimates ----
-    p <- anthro.02 |>
-      mw_estimate_prevalence_screening(
-        muac = muac,
-        age = age,
-        oedema = oedema,
-        province
-      )
 
-    ### Observed estimates ----
-    gam_n <- 61
-    gam_p <- 5.9
-    sam_n <- 19
-    sam_p <- 1.8
-    mam_n <- 42
-    mam_p <- 4.1
+## ---- mw_estimate_prevalence_screening() -------------------------------------
 
-    ### Tests ----
-    testthat::expect_s3_class(p, "tbl_df")
-    testthat::expect_equal(ncol(p), 8)
-    testthat::expect_equal(nrow(p), 2)
-    testthat::expect_true(
-      all(c("province", "gam_n", "gam_p", "sam_n", "sam_p", "mam_n", "mam_p") %in% names(p))
-    )
-    testthat::expect_equal(p[[2]][1], gam_n)
-    testthat::expect_equal(round(p[[3]][1] * 100, 1), gam_p)
-    testthat::expect_equal(p[[4]][1], sam_n)
-    testthat::expect_equal(round(p[[5]][1] * 100, 1), sam_p)
-    testthat::expect_equal(p[[6]][1], mam_n)
-    testthat::expect_equal(round(p[[7]][1] * 100, 1), mam_p)
-  }
-)
 
-## When grouping variables are not supplied
+### When grouping variables are not supplied
 testthat::test_that(
   "mw_estimate_prevalence_screening() works OK when grouping variables are not supplied",
   {
-    ### Get estimates ----
-    p <- anthro.02 |>
-      mw_estimate_prevalence_screening(
-        muac = muac,
-        age = age,
-        oedema = oedema
-      )
+    #### Get estimates ----
+    p <- mw_estimate_prevalence_screening(anthro.02, muac, age, oedema)
 
-    ### Observed estimates ----
+    #### Observed estimates ----
     gam_n <- 118
     gam_p <- 5.4
     sam_n <- 29
@@ -266,7 +202,7 @@ testthat::test_that(
     mam_p <- 4.1
 
 
-    ### Tests ----
+    #### Tests ----
     testthat::expect_s3_class(p, "tbl_df")
     testthat::expect_equal(ncol(p), 7)
     testthat::expect_equal(nrow(p), 1)
@@ -282,138 +218,79 @@ testthat::test_that(
   }
 )
 
-## When an age-weighting approach is applied ----
-testthat::test_that(
-  "mw_estimate_prevalence_screening() applies age-weighting correctly",
-  {
-    ## Observed results ----
-    p <- anthro.04 |>
-      subset(province == "Province 2") |>
-      mw_estimate_prevalence_screening(muac = muac, age = age, oedema = NULL)
 
-    ## Tests ----
-    testthat::expect_s3_class(p, "tbl")
-    testthat::expect_equal(round(p[[3]][1] * 100, 1), 8.6)
-    testthat::expect_equal(round(p[[1]][1] * 100, 1), 1.5)
-    testthat::expect_equal(round(p[[2]][1] * 100, 1), 7.1)
-  }
-)
-
-## When used on a multiple-area dataset ----
+### When used on a multiple-area dataset ----
 testthat::test_that(
   "mw_estimate_prevalence_screening() works well on a multiple-area dataset with
     different categories of analysis_approach",
   {
-    ### Get the prevalence estimates ----
+    #### Get the prevalence estimates ----
     p <- anthro.04 |>
-      mw_estimate_prevalence_screening(
-        muac = muac, 
-        age = age,
-        oedema = oedema, 
-        province
-      )
+      mw_wrangle_age(age = age) |> 
+      mw_wrangle_muac(sex, muac, age, FALSE, TRUE, "cm") |> 
+      mutate(muac = recode_muac(muac, "mm")) |> 
+      mw_estimate_prevalence_screening(muac, age, oedema, analysis_unit)
 
     columns_to_check <- c("gam_n", "gam_p", "sam_n", "sam_p", "mam_n", "mam_p", "N")
 
-    ### test ----
-    testthat::expect_vector(select(p, !province), size = 3, ncol(8))
+    #### test ----
+    testthat::expect_vector(select(p, !analysis_unit), size = 3, ncol(8))
     testthat::expect_s3_class(p, "tbl")
     testthat::expect_false(all(sapply(p[2, ][columns_to_check], \(.) all(is.na(.)))))
 
-    ### Province 2 ----
+    #### Unit B: age-weighted prevalence ----
     testthat::expect_true(is.na(p[2, 2][[1]]))
-    testthat::expect_equal(round(p[2, 3][[1]] * 100, 1), 8.6)
+    testthat::expect_equal(round(p[2, 3][[1]] * 100, 1), 12.1)
     testthat::expect_true(is.na(p[2, 4][[1]]))
-    testthat::expect_equal(round(p[2, 5][[1]] * 100, 1), 1.5)
+    testthat::expect_equal(round(p[2, 5][[1]] * 100, 1), 2.9)
     testthat::expect_true(is.na(p[2, 6][[1]]))
-    testthat::expect_equal(round(p[2, 7][[1]] * 100, 1), 7.1)
+    testthat::expect_equal(round(p[2, 7][[1]] * 100, 1), 9.2)
+    testthat::expect_equal(p[2, 8][[1]], 1354)
 
-    ### Province 3 ----
-    testthat::expect_equal(round(p[3, 3][[1]] * 100, 1), 11.1)
-    testthat::expect_equal(round(p[3, 5][[1]] * 100, 1), 2.8)
-    testthat::expect_equal(round(p[3, 7][[1]] * 100, 1), 8.3)
+    #### Unit C: Age ratio is problematic but prop >= 0.66 (non-age-weighted p) ----
+    testthat::expect_equal(round(p[3, 3][[1]] * 100, 1), 9.1)
+    testthat::expect_equal(round(p[3, 5][[1]] * 100, 1), 2.4)
+    testthat::expect_equal(round(p[3, 7][[1]] * 100, 1), 6.7)
+    testthat::expect_equal(p[3, 8][[1]], 209)
   }
 )
 
-# Test check: mw_estimate_prevalence_screening2() ----
-testthat::test_that(
-  "mw_estimate_prevalence_screening2() works as expected when grouping vars are not supplied",
-  {
-    ## Observed results ----
-    p <- anthro.01 |>
-      mutate(age_cat = ifelse(age < 24, "6-23", "24-59")) |>
-      mw_wrangle_muac(
-        sex = sex,
-        .recode_sex = TRUE,
-        muac = muac
-      ) |>
-      mw_estimate_prevalence_screening2(
-        age_cat = age_cat,
-        muac = muac
-      )
 
-    ## Tests ----
-    testthat::expect_s3_class(p, "tbl_df")
-    testthat::expect_equal(round(p[[2]] * 100, 1), 2.9)
-  }
-)
+## ---- mw_estimate_prevalence_screening2() ------------------------------------
 
+
+### When grouping variables are given ----
 testthat::test_that(
   "mw_estimate_prevalence_screening2() works as expected when grouping vars are specified",
   {
-    ## Observed results ----
-    p <- anthro.01 |>
-      mutate(age_cat = ifelse(age < 24, "6-23", "24-59")) |>
-      mw_wrangle_muac(
-        sex = sex,
-        .recode_sex = TRUE,
-        .recode_muac = FALSE,
-        .to = "none",
-        muac = muac
-      ) |>
-      mw_estimate_prevalence_screening2(
-        age_cat = age_cat,
-        muac = muac,
-        oedema = NULL,
-        area
-      )
-
-    ## Tests ----
-    testthat::expect_s3_class(p, "tbl_df")
-    testthat::expect_equal(round(p[[3]][2] * 100, 2), 3.22)
-    testthat::expect_equal(names(p[1]), "area")
-  }
-)
-
-## Test-check: mw_estimate_prevalence_screening2() ----
-testthat::test_that(
-  "mw_estimate_prevalence_screening2() returns correct estimates for weighted analysis",
-  {
-    ### Get the prevalence estimates ----
+    #### Observed results ----
     p <- anthro.04 |>
       mutate(age_cat = ifelse(age < 24, "6-23", "24-59")) |>
-      mw_wrangle_muac(
-        muac = muac,
-        .recode_muac = FALSE,
-        .to = "none",
-        sex = sex,
-        .recode_sex = FALSE
-      ) |>
+      mw_wrangle_muac(sex = sex, .recode_sex = TRUE, muac = muac) |>
       mw_estimate_prevalence_screening2(
         age_cat = age_cat,
-        muac = muac,
-        oedema = NULL,
-        province
+        muac = muac, 
+        oedema = oedema, 
+        analysis_unit
       )
 
-
-    ### Tests ----
+    #### Tests ----
     testthat::expect_s3_class(p, "tbl_df")
-    testthat::expect_equal(round(p[[3]][1] * 100, 1), expected = 10.5)
-    testthat::expect_equal(p[[2]][1], expected = 135)
-    testthat::expect_equal(p[[4]][1], expected = 19)
-    testthat::expect_equal(p[[6]][1], expected = 116)
-    testthat::expect_true(is.na(p[[2]][2]))
-    testthat::expect_equal(round(p[[3]][3] * 100, 1), expected = 11.9)
+
+    #### Unit B: age-weighted prevalence ----
+    testthat::expect_true(is.na(p[2, 2][[1]]))
+    testthat::expect_equal(round(p[2, 3][[1]] * 100, 1), 12.6)
+    testthat::expect_true(is.na(p[2, 4][[1]]))
+    testthat::expect_equal(round(p[2, 5][[1]] * 100, 1), 3.4)
+    testthat::expect_true(is.na(p[2, 6][[1]]))
+    testthat::expect_equal(round(p[2, 7][[1]] * 100, 1), 9.1)
+    testthat::expect_equal(p[2, 8][[1]], 1365)
+
+    #### Unit C: Age ratio is problematic but prop >= 0.66 (non-age-weighted p) ----
+    testthat::expect_equal(round(p[3, 3][[1]] * 100, 1), 9.5)
+    testthat::expect_equal(round(p[3, 5][[1]] * 100, 1), 2.9)
+    testthat::expect_equal(round(p[3, 7][[1]] * 100, 1), 6.7)
+    testthat::expect_equal(p[3, 8][[1]], 210)
   }
 )
+